fix: extract protocol setup and load calculation
