+++ user-files/simulations/infrastructure/RegistrySimulation.scala
@@ -1,6 +1,5 @@
 package infrastructure
 
-import io.gatling.http.Predef._
 import io.gatling.core.Predef._
 import io.gatling.core.structure.ChainBuilder
 import io.gatling.core.structure.PopulationBuilder
@@ -9,39 +8,21 @@ import scala.concurrent.duration._
 import scala.math.{min, max}
 
 abstract class RegistrySimulation extends Simulation {
-  
-  private val loadTestApiKey = System.getProperty("api_key")
-
-  val httpProtocol = http
-    .baseUrl(System.getProperty("base_url"))
-    .header("x-api-key", loadTestApiKey)
-    .warmUp(System.getProperty("warmup_url"))
-    .acceptHeader("application/json, text/javascript, */*; q=0.5")
-    .acceptLanguageHeader("en-US,en;q=0.9,nl;q=0.8")
-    .acceptEncodingHeader("gzip, deflate, br")
-    .userAgentHeader("Basisregisters Vlaanderen Gatling - Load Test Simulation")
-    .doNotTrackHeader("1")
-    .disableCaching
-
+    
   def setUp(
     name: String, 
     load: TrafficLoadConfiguration, 
     possibilities: List[Possibility]): SetUp = {          
-    
-    val numberOfUsers = atLeast1(load.numberOfUsers)
-    val userIncrementation = atLeast1(atMost20(numberOfUsers / 5))
-    val numberOfLevels = atLeast1(numberOfUsers / userIncrementation) 
-    println("numberOfUsers: " + numberOfUsers + ", incrementNumberOfUsersPerCycle: " + userIncrementation + ", NumberOfCycles: " + numberOfLevels)
 
     setUp(
       scenario(name)
         .exec(createSwitch(possibilities, load))
         .inject(
-          incrementUsersPerSec(userIncrementation)
-            .times(numberOfLevels)
+          incrementUsersPerSec(load.incrementUsersPerCycleBy)
+            .times(load.numberOfCycles)
             .eachLevelLasting(10 seconds)
-            .startingFrom(userIncrementation)
-        ).protocols(httpProtocol)
+            .startingFrom(load.incrementUsersPerCycleBy)
+        ).protocols(Protocols.default)
     )
   }
